Move recipemaps to separated class
