+++ src/main/java/gtPlusPlus/xmod/gregtech/common/tileentities/generators/GregtechMetaTileEntityRocketFuelGenerator.java
@@ -12,9 +12,9 @@ import gregtech.api.interfaces.ITexture;
 import gregtech.api.interfaces.tileentity.IGregTechTileEntity;
 import gregtech.api.metatileentity.MetaTileEntity;
 import gregtech.api.objects.GT_RenderedTexture;
-import gregtech.api.util.GTPP_Recipe;
 import gregtech.api.util.GT_ModHandler;
 import gregtech.api.util.GT_Recipe;
+import gtPlusPlus.api.recipe.GTPPRecipeMaps;
 import gtPlusPlus.xmod.gregtech.api.metatileentity.implementations.base.generators.GregtechRocketFuelGeneratorBase;
 import gtPlusPlus.xmod.gregtech.common.blocks.textures.TexturesGtBlock;
 
@@ -50,7 +50,7 @@ public class GregtechMetaTileEntityRocketFuelGenerator extends GregtechRocketFue
 
     @Override
     public GT_Recipe.GT_Recipe_Map getRecipes() {
-        return GTPP_Recipe.GTPP_Recipe_Map.sRocketFuels;
+        return GTPPRecipeMaps.sRocketFuels;
     }
 
     @Override
