+++ src/main/java/gtPlusPlus/xmod/gregtech/common/tileentities/machines/multi/processing/GregtechMetaTileEntity_NuclearSaltProcessingPlant.java
@@ -29,6 +29,7 @@ import gregtech.api.metatileentity.MetaTileEntity;
 import gregtech.api.util.GT_Multiblock_Tooltip_Builder;
 import gregtech.api.util.GT_Recipe;
 import gregtech.api.util.GT_Utility;
+import gtPlusPlus.api.recipe.GTPPRecipeMaps;
 import gtPlusPlus.core.block.ModBlocks;
 import gtPlusPlus.core.lib.CORE;
 import gtPlusPlus.xmod.gregtech.api.metatileentity.implementations.base.GregtechMeta_MultiBlockBase;
@@ -185,7 +186,7 @@ public class GregtechMetaTileEntity_NuclearSaltProcessingPlant
 
     @Override
     public GT_Recipe.GT_Recipe_Map getRecipeMap() {
-        return gregtech.api.util.GTPP_Recipe.GTPP_Recipe_Map.sNuclearSaltProcessingPlantRecipes;
+        return GTPPRecipeMaps.sNuclearSaltProcessingPlantRecipes;
     }
 
     @Override
