+++ src/main/java/gtPlusPlus/xmod/gregtech/loaders/misc/AddCustomMachineToPA.java
@@ -1,14 +1,13 @@
 package gtPlusPlus.xmod.gregtech.loaders.misc;
 
-import gregtech.api.util.GTPP_Recipe;
 import gregtech.api.util.GT_ProcessingArray_Manager;
+import gtPlusPlus.api.recipe.GTPPRecipeMaps;
 
 public class AddCustomMachineToPA {
 
     public static void register() {
 
         // Simple Washers
-        GT_ProcessingArray_Manager
-                .addRecipeMapToPA("simplewasher.01", GTPP_Recipe.GTPP_Recipe_Map.sSimpleWasherRecipes);
+        GT_ProcessingArray_Manager.addRecipeMapToPA("simplewasher.01", GTPPRecipeMaps.sSimpleWasherRecipes);
     }
 }
