+++ src/main/java/gtPlusPlus/xmod/gregtech/registration/gregtech/GregtechIndustrialMassFabricator.java
@@ -5,9 +5,9 @@ import net.minecraftforge.fluids.FluidStack;
 
 import gregtech.api.enums.GT_Values;
 import gregtech.api.enums.Materials;
-import gregtech.api.util.GTPP_Recipe;
 import gregtech.api.util.GT_Recipe;
 import gtPlusPlus.api.objects.Logger;
+import gtPlusPlus.api.recipe.GTPPRecipeMaps;
 import gtPlusPlus.core.lib.CORE;
 import gtPlusPlus.core.recipe.common.CI;
 import gtPlusPlus.core.util.minecraft.ItemUtils;
@@ -62,8 +62,8 @@ public class GregtechIndustrialMassFabricator {
                 32,
                 0);
 
-        GTPP_Recipe.GTPP_Recipe_Map.sMatterFab2Recipes.add(UUA_From_Scrap);
-        GTPP_Recipe.GTPP_Recipe_Map.sMatterFab2Recipes.add(UUA_From_ScrapBoxes);
+        GTPPRecipeMaps.sMatterFab2Recipes.add(UUA_From_Scrap);
+        GTPPRecipeMaps.sMatterFab2Recipes.add(UUA_From_ScrapBoxes);
 
         // Basic UUM
         GT_Recipe generateUUM_LV = new GT_Recipe(
@@ -92,7 +92,7 @@ public class GregtechIndustrialMassFabricator {
                 0);
 
         // Advanced UUM
-        GTPP_Recipe.GTPP_Recipe_Map.sMatterFab2Recipes.add(
+        GTPPRecipeMaps.sMatterFab2Recipes.add(
                 new GT_Recipe(
                         false,
                         new ItemStack[] { CI.getNumberedCircuit(3) },
@@ -106,7 +106,7 @@ public class GregtechIndustrialMassFabricator {
                         0));
 
         // Advanced UUM
-        GTPP_Recipe.GTPP_Recipe_Map.sMatterFab2Recipes.add(
+        GTPPRecipeMaps.sMatterFab2Recipes.add(
                 new GT_Recipe(
                         false,
                         new ItemStack[] { CI.getNumberedCircuit(4) },
@@ -119,12 +119,11 @@ public class GregtechIndustrialMassFabricator {
                         65536,
                         0));
 
-        GTPP_Recipe.GTPP_Recipe_Map.sMatterFab2Recipes.add(generateUUM_LV);
-        GTPP_Recipe.GTPP_Recipe_Map.sMatterFab2Recipes.add(generateUUMFromUUA_LV);
+        GTPPRecipeMaps.sMatterFab2Recipes.add(generateUUM_LV);
+        GTPPRecipeMaps.sMatterFab2Recipes.add(generateUUMFromUUA_LV);
 
         Logger.INFO(
-                "Generated " + GTPP_Recipe.GTPP_Recipe_Map.sMatterFab2Recipes.mRecipeList.size()
-                        + " Matter Fabricator recipes.");
+                "Generated " + GTPPRecipeMaps.sMatterFab2Recipes.mRecipeList.size() + " Matter Fabricator recipes.");
     }
 
     public static ItemStack getScrapPile() {
