+++ src/main/java/gtPlusPlus/xmod/gregtech/registration/gregtech/GregtechSolarTower.java
@@ -2,8 +2,8 @@ package gtPlusPlus.xmod.gregtech.registration.gregtech;
 
 import net.minecraftforge.fluids.FluidStack;
 
-import gregtech.api.util.GTPP_Recipe;
 import gtPlusPlus.api.objects.Logger;
+import gtPlusPlus.api.recipe.GTPPRecipeMaps;
 import gtPlusPlus.core.material.MISC_MATERIALS;
 import gtPlusPlus.xmod.gregtech.api.enums.GregtechItemList;
 import gtPlusPlus.xmod.gregtech.common.tileentities.machines.multi.production.GregtechMetaTileEntity_SolarTower;
@@ -31,7 +31,7 @@ public class GregtechSolarTower {
                         0).getStackForm(1L));
 
         // NEI recipe
-        GTPP_Recipe.GTPP_Recipe_Map.sSolarTowerRecipes.addRecipe(
+        GTPPRecipeMaps.sSolarTowerRecipes.addRecipe(
                 false,
                 null,
                 null,
